Постман побеждён, добавил диаграмму БД, мелкие правки

+++ src/main/java/ru/practicum/shareit/item/repository/CommentRepository.java
@@ -5,8 +5,10 @@ import ru.practicum.shareit.item.model.Comment;
 
 import java.util.List;
 
-public interface CommentRepository extends JpaRepository<Comment, Long> {
+public interface CommentRepository extends JpaRepository<Comment, Integer> {
+
     List<Comment> findAllByItemId(int itemId);
 
     List<Comment> findAllByItemIdIn(List<Integer> itemIds);
+
 }
